+++ brickrail-gui/devices/devices.gd
@@ -37,6 +37,9 @@ func load(struct):
 	
 	for train_data in struct.trains:
 		var _train = add_train(train_data.name)
+		if "storage" in train_data:
+			for key in _train.hub.storage:
+				_train.hub.store_value(int(key), int(train_data.storage[str(key)]))
 		# train.load(train_data)
 	
 	for controller_data in struct.controllers: